chore: run tests sequentially not parallel & more functions
