+++ __test__/kross-client/base.tsx
@@ -1,0 +1,57 @@
+import { QueryClient, QueryClientProvider } from 'react-query';
+import { KrossClient } from '../../src/kross-client';
+import React from 'react';
+import { act, renderHook, waitFor } from '@testing-library/react';
+
+export const base = () => {
+    let client: KrossClient;
+    const baseURL = 'https://olive-dev.kross.kr';
+    const accessId = 'XLD7UY9GETOK7TPY';
+    const secretKey = 'yLbVRHGgwT5c22ndOVT2';
+    const queryClient = new QueryClient({
+      defaultOptions: {
+        queries: {
+          // ✅ turns retries off
+          retry: false,
+        },
+      },
+    });
+    const wrapper = ({ children }) => (
+      <QueryClientProvider client={queryClient}>{children}</QueryClientProvider>
+    );
+    beforeAll(() => {
+      client = new KrossClient({
+        baseURL,
+        accessId,
+        secretKey,
+        adapter: require('axios/lib/adapters/http'),
+      });
+    });
+  
+    it('gets authToken and refreshToken', async () => {
+      const { useLogin } = client.useAuthHooks();
+      const { result } = renderHook(() => useLogin(), {
+        wrapper,
+      });
+      await act(async () => {
+        await result.current.mutateAsync({
+          keyid: 'mad@kross.kr',
+          password: 'Kross123!',
+        });
+      });
+      await waitFor(() => expect(result.current.isSuccess).toBe(true));
+      expect(result.current.data).toBeDefined();
+    }, 30000);
+  
+    it.skip('updates authToken with refreshToken', async () => {
+      const { updateAuthToken } = client.useAuthHooks();
+      const { result } = renderHook(() => updateAuthToken(), {
+        wrapper,
+      });
+      await act(async () => {
+        await result.current.refetch();
+      });
+      await waitFor(() => expect(result.current.isSuccess).toBe(true));
+      expect(result.current.data).toBeDefined();
+    }, 30000);  
+}
