+++ src/kross-client/loans.ts
@@ -44,6 +44,19 @@ export class Loans extends KrossClientBase {
       }
     );
   }
+
+  async recentFundingItem() {
+    const recentProduct = await this.loanData({
+      fields: 'id,name,fund_amount,payment_date,due_date,category,interest_rate,investor_fee_rate,state',
+      sort_by: 'id.desc',
+      query: {
+        state: 'funding',
+      },
+      limit: '3',
+    });
+
+    return recentProduct;
+  }
   useLoanHooks() {
     return {
       loanConfigs: (loansQueryDto: LoansQueryDto) => {
@@ -69,6 +82,12 @@ export class Loans extends KrossClientBase {
           queryKey: 'loanData',
           queryFn: async () => await this.loanData(loansQueryDto),
         });
+      },
+      recentFundingItem: () => {
+        return useQuery({
+          queryKey: 'recentFundingItem',
+          queryFn: async () => await this.recentFundingItem()
+        })
       },
     };
   }
